pyproject.toml
[project]
name = "logbless"
version = "0.1.0"
description = "Simple logs for any app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]


[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]


[project.scripts]
logbless = "logbless.cli:cli"
